requirements.txt
matplotlib==3.2.1
requests==2.18.4
numpy==1.13.3
jittor==1.2.2.13
tqdm==4.46.0
scipy==1.6.0
Pillow==8.1.0
